Eventmodel conversion to typescript
